+++ core/src/main/java/nl/gjorgdy/pl3xmarkers/core/interfaces/api/IWorldApi.java
@@ -1,9 +1,11 @@
 package nl.gjorgdy.pl3xmarkers.core.interfaces.api;
 
 import nl.gjorgdy.pl3xmarkers.core.layers.primitive.MarkerLayer;
+import org.jspecify.annotations.Nullable;
 
 public interface IWorldApi {
 
+	@Nullable
 	<T extends MarkerLayer> T getLayer(Class<T> layerClass, String layerKey);
 
 }