+++ fabric/src/main/java/nl/gjorgdy/pl3xmarkers/fabric/mixin/PersistentEntitySectionManagerMixin.java
@@ -17,10 +17,13 @@ public class PersistentEntitySectionManagerMixin<T extends EntityAccess> {
 	@Inject(method = "addEntity(Lnet/minecraft/world/level/entity/EntityAccess;Z)Z", at = @At("HEAD"))
 	public void onEntityAdd(T entity, boolean loaded, CallbackInfoReturnable<Boolean> cir) {
 		if (entity instanceof LightningBolt lightning) {
-			Pl3xMarkersCore.api()
+			var markerLayer = Pl3xMarkersCore.api()
 					.getWorld(lightning.level().dimension().identifier().toString())
-					.getLayer(LightningMarkerLayer.class, Layers.Keys.LIGHTNING)
-					.show(lightning.getBlockX(), lightning.getBlockY(), lightning.getBlockZ());
+					.getLayer(LightningMarkerLayer.class, Layers.Keys.LIGHTNING);
+			if (markerLayer == null) {
+				return;
+			}
+			markerLayer.show(lightning.getBlockX(), lightning.getBlockY(), lightning.getBlockZ());
 		}
 	}
 }