+++ changelog.md
@@ -3,4 +3,5 @@
 
 - Fixed sign markers not being removed correctly when the sign is removed from the world.
 - Fixed migration happening on every startup creating a lot of folders.
+- Fixed a crash caused by disabling some of the marker layers.
 - Fixed feedback sounds not playing on Fabric.